+++ ccv/models/__init__.py
@@ -1,0 +1,3 @@
+from ccv.models import (
+    base, personal_information, education, activity, contribution, employment, user_profile, recognitions
+)
